requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements/dev.txt requirements/dev.in
#
-e git+https://github.com/frankhood/django-self-aware-model.git@0.1.0#egg=self-aware-model
    # via
    #   -r requirements/common.in
    #   -r requirements/dev.in
-e git+https://github.com/frankhood/django-templatetag-toolkit.git@0.1.1#egg=templatetag-toolkit
    # via -r requirements/dev.in
arrow==1.2.2
    # via jinja2-time
asgiref==3.5.2
    # via django
asttokens==2.0.5
    # via stack-data
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
attrs==21.4.0
    # via
    #   outcome
    #   pytest
    #   trio
backcall==0.2.0
    # via ipython
binaryornot==0.4.4
    # via cookiecutter
certifi==2022.5.18.1
    # via
    #   requests
    #   urllib3
cffi==1.15.0
    # via cryptography
cfgv==3.3.1
    # via pre-commit
chardet==4.0.0
    # via binaryornot
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   cookiecutter
    #   pip-tools
cookiecutter==1.7.3
    # via -r requirements/dev.in
coverage==5.3.1
    # via
    #   -r requirements/dev.in
    #   -r requirements/tests.in
    #   django-coverage-plugin
cryptography==37.0.2
    # via
    #   pyopenssl
    #   urllib3
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
distlib==0.3.4
    # via virtualenv
dj-database-url==0.5.0
    # via -r requirements/common.in
dj-email-url==1.0.5
    # via -r requirements/common.in
django==3.2.13
    # via
    #   -r requirements/common.in
    #   django-classy-tags
    #   django-configurations
    #   django-extensions
    #   django-prettyjson
    #   djangorestframework
    #   model-bakery
    #   self-aware-model
    #   templatetag-toolkit
django-classy-tags==3.0.1
    # via
    #   -r requirements/common.in
    #   templatetag-toolkit
django-configurations==2.3.2
    # via -r requirements/common.in
django-coverage-plugin==2.0.3
    # via -r requirements/tests.in
django-extensions==3.1.5
    # via
    #   -r requirements/common.in
    #   -r requirements/dev.in
django-prettyjson==0.4.1
    # via -r requirements/common.in
django-widget-tweaks==1.4.12
    # via -r requirements/common.in
djangorestframework==3.13.1
    # via -r requirements/common.in
executing==0.8.3
    # via stack-data
factory-boy==3.2.1
    # via -r requirements/tests.in
faker==13.12.0
    # via
    #   -r requirements/tests.in
    #   factory-boy
feedparser==6.0.10
    # via -r requirements/tests.in
filelock==3.7.0
    # via
    #   tox
    #   virtualenv
flake8==4.0.1
    # via -r requirements/dev.in
freezegun==1.2.1
    # via -r requirements/tests.in
h11==0.13.0
    # via wsproto
identify==2.5.1
    # via pre-commit
idna==3.3
    # via
    #   requests
    #   trio
    #   urllib3
iniconfig==1.1.1
    # via pytest
invoke==1.7.1
    # via -r requirements/dev.in
ipdb==0.13.9
    # via -r requirements/dev.in
ipython==8.4.0
    # via
    #   -r requirements/common.in
    #   ipdb
isort==5.10.1
    # via -r requirements/dev.in
jedi==0.17.2
    # via
    #   -r requirements/dev.in
    #   ipython
jinja2==3.1.2
    # via
    #   cookiecutter
    #   jinja2-time
jinja2-time==0.2.0
    # via cookiecutter
markupsafe==2.1.1
    # via jinja2
matplotlib-inline==0.1.3
    # via ipython
mccabe==0.6.1
    # via flake8
model-bakery==1.5.0
    # via -r requirements/tests.in
nodeenv==1.6.0
    # via pre-commit
outcome==1.1.0
    # via trio
packaging==21.3
    # via
    #   pytest
    #   tox
parso==0.7.1
    # via jedi
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.1.1
    # via -r requirements/common.in
pip-tools==6.5.1
    # via -r requirements/common.in
platformdirs==2.5.2
    # via virtualenv
pluggy==1.0.0
    # via
    #   pytest
    #   tox
poyo==0.5.0
    # via cookiecutter
pre-commit==2.9.3
    # via -r requirements/tests.in
prompt-toolkit==3.0.29
    # via ipython
psycopg2-binary==2.9.3
    # via -r requirements/common.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via
    #   pytest
    #   tox
pycodestyle==2.8.0
    # via flake8
pycparser==2.21
    # via cffi
pydantic==1.9.1
    # via -r requirements/common.in
pydot==1.4.2
    # via -r requirements/dev.in
pyflakes==2.4.0
    # via flake8
pygments==2.12.0
    # via ipython
pyopenssl==22.0.0
    # via urllib3
pyparsing==3.0.9
    # via
    #   packaging
    #   pydot
pysocks==1.7.1
    # via urllib3
pytest==7.1.2
    # via pytest-django
pytest-django==4.1.0
    # via -r requirements/dev.in
python-dateutil==2.8.2
    # via
    #   -r requirements/common.in
    #   arrow
    #   faker
    #   freezegun
python-dotenv==0.20.0
    # via -r requirements/common.in
python-levenshtein==0.12.2
    # via -r requirements/common.in
python-slugify==6.1.2
    # via cookiecutter
pytz==2022.1
    # via
    #   -r requirements/common.in
    #   django
    #   djangorestframework
pyupgrade==2.7.4
    # via -r requirements/dev.in
pyyaml==6.0
    # via pre-commit
raven==6.10.0
    # via -r requirements/common.in
requests==2.27.1
    # via
    #   -r requirements/common.in
    #   cookiecutter
rope==0.18.0
    # via -r requirements/dev.in
selenium==4.2.0
    # via -r requirements/tests.in
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   -r requirements/common.in
    #   asttokens
    #   cookiecutter
    #   django-coverage-plugin
    #   django-prettyjson
    #   python-dateutil
    #   tox
    #   virtualenv
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via trio
sqlparse==0.4.2
    # via django
stack-data==0.2.0
    # via ipython
standardjson==0.3.1
    # via django-prettyjson
text-unidecode==1.3
    # via
    #   -r requirements/tests.in
    #   python-slugify
tokenize-rt==4.2.1
    # via pyupgrade
toml==0.10.2
    # via
    #   ipdb
    #   pre-commit
    #   tox
tomli==2.0.1
    # via
    #   pep517
    #   pytest
tox==3.20.1
    # via -r requirements/tests.in
traitlets==5.2.1.post0
    # via
    #   ipython
    #   matplotlib-inline
trio==0.20.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
typing-extensions==4.2.0
    # via pydantic
urllib3[secure,socks]==1.26.9
    # via
    #   requests
    #   selenium
uwsgi==2.0.20
    # via -r requirements/dev.in
virtualenv==20.14.1
    # via
    #   pre-commit
    #   tox
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.1
    # via pip-tools
wsproto==1.1.0
    # via trio-websocket

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
